move ground type declarations to Proof.Instances
