+++ src/Pate/SimulatorRegisters.hs
@@ -15,8 +15,7 @@ module Pate.SimulatorRegisters (
   MacawRegVar(..),
   MacawRegEntry(..),
   macawRegEntry,
-  ptrToEntry,
-  ValidMacawType
+  ptrToEntry
   ) where
 
 import qualified Data.Macaw.Symbolic as MS
@@ -44,12 +43,10 @@ type family CrucBaseTypes (tp :: CT.CrucibleType) :: Ctx.Ctx WI.BaseType where
   CrucBaseTypes CT.BoolType = (Ctx.EmptyCtx Ctx.::> WT.BaseBoolType)
   CrucBaseTypes (CT.StructType Ctx.EmptyCtx) = Ctx.EmptyCtx
 
-type ValidMacawType tp = (Eq (PT.ConcreteValue (MS.ToCrucibleType tp)), Show (PT.ConcreteValue (MS.ToCrucibleType tp)))
-
 -- | This is an analog of the Crucible 'CS.RegEntry' type in terms of the macaw
 -- type system
 data MacawRegEntry sym (tp :: MT.Type) where
-  MacawRegEntry :: ValidMacawType tp =>
+  MacawRegEntry ::
     { macawRegRepr :: CT.TypeRepr (MS.ToCrucibleType tp)
     , macawRegValue :: CS.RegValue sym (MS.ToCrucibleType tp)
     } ->
@@ -75,11 +72,7 @@ instance PC.ShowF (WI.SymExpr sym) => Show (MacawRegEntry sym tp) where
     _ -> "macawRegEntry: unsupported"
 
 macawRegEntry :: CS.RegEntry sym (MS.ToCrucibleType tp) -> MacawRegEntry sym tp
-macawRegEntry (CS.RegEntry repr v) = case repr of
-  CLM.LLVMPointerRepr{} -> MacawRegEntry repr v
-  CT.BoolRepr -> MacawRegEntry repr v
-  CT.StructRepr Ctx.Empty -> MacawRegEntry repr v
-  _ -> error ("macawRegEntry: unsupported macaw type " ++ show repr)
+macawRegEntry (CS.RegEntry repr v) = MacawRegEntry repr v
 
 ptrToEntry ::
   WI.IsExprBuilder sym =>
